[WIP] Store open timeslots in a separate table
In order to more efficiently find offers with timeslots that overlap
a bid, we maintain a timeslots table in the database. When an offer is
first created, we add a single timelot that matches the start/end time
of the offer. Whena a contract is created, we remove the existing
timeslot and replace it with (at most) two new entries: (1) the
available time *before* the contract starts and (2) the available time
*after* the contract starts.

This should allow us to find open timeslots with a single query on a
join of the offers and timeslots tables.